Fix same team flashes
